Fixes breaking change with missing NHibernateStorageContext when using outbox
