+++ src/NServiceBus.NHibernate/NServiceBus.NHibernate.csproj
@@ -123,8 +123,9 @@
     <Compile Include="SagaPersisters\SagaPersister.cs" />
     <Compile Include="SharedSession\SessionFactoryProvider.cs" />
     <Compile Include="SharedSession\SharedMappings.cs" />
-    <Compile Include="SharedSession\TransportIntegrationFeature.cs" />
+    <Compile Include="SqlServerTransport\DisabledSharedConnectionByDefault.cs" />
     <Compile Include="SharedSession\SharedConnectionStorageSessionProvider.cs" />
+    <Compile Include="SqlServerTransport\ValidateAmbientTransactionOrOutbox.cs" />
     <Compile Include="Subscriptions\CachedSubscriptionPersister.cs" />
     <Compile Include="Subscriptions\NHibernateSubscriptionStorage.cs" />
     <Compile Include="Subscriptions\SubscriptionConfig.cs" />
